test and refactor

+++ vae/encoder_decoder.py
@@ -1,5 +1,3 @@
-
-import math
 import torch
 import torch.nn as nn
 import numpy as np
@@ -9,7 +7,7 @@ from basicsr.ops.dcn import ModulatedDeformConvPack, modulated_deform_conv
 
 
 class Encoder(nn.Module):
-    def __init__(self, ch=128, out_ch=3, ch_mult=[1,2,4], num_res_blocks=2,
+    def __init__(self, ch=128, out_ch=3, ch_mult=[1, 2, 4], num_res_blocks=2,
                  attn_resolutions=[], dropout=0.0, resamp_with_conv=True, in_channels=3,
                  resolution=256, z_channels=3, double_z=True, use_linear_attn=False, attn_type="vanilla"):
         super().__init__()
@@ -29,14 +27,14 @@ class Encoder(nn.Module):
                                        padding=1)
 
         curr_res = resolution
-        in_ch_mult = (1,)+tuple(ch_mult)
+        in_ch_mult = (1,) + tuple(ch_mult)
         self.in_ch_mult = in_ch_mult
         self.down = nn.ModuleList()
         for i_level in range(self.num_resolutions):
             block = nn.ModuleList()
             attn = nn.ModuleList()
-            block_in = ch*in_ch_mult[i_level]
-            block_out = ch*ch_mult[i_level]
+            block_in = ch * in_ch_mult[i_level]
+            block_out = ch * ch_mult[i_level]
             for i_block in range(self.num_res_blocks):
                 block.append(ResnetBlock(in_channels=block_in,
                                          out_channels=block_out,
@@ -48,7 +46,7 @@ class Encoder(nn.Module):
             down = nn.Module()
             down.block = block
             down.attn = attn
-            if i_level != self.num_resolutions-1:
+            if i_level != self.num_resolutions - 1:
                 down.downsample = Downsample(block_in, resamp_with_conv)
                 curr_res = curr_res // 2
             self.down.append(down)
@@ -68,7 +66,7 @@ class Encoder(nn.Module):
         # end
         self.norm_out = Normalize(block_in)
         self.conv_out = torch.nn.Conv2d(block_in,
-                                        2*z_channels if double_z else z_channels,
+                                        2 * z_channels if double_z else z_channels,
                                         kernel_size=3,
                                         stride=1,
                                         padding=1)
@@ -85,7 +83,7 @@ class Encoder(nn.Module):
                 if len(self.down[i_level].attn) > 0:
                     h = self.down[i_level].attn[i_block](h)
                 hs.append(h)
-            if i_level != self.num_resolutions-1:
+            if i_level != self.num_resolutions - 1:
                 enc_feat.append(hs[-1])
                 hs.append(self.down[i_level].downsample(hs[-1]))
 
@@ -106,9 +104,9 @@ class Encoder(nn.Module):
 
 
 class Decoder(nn.Module):
-    def __init__(self, ch, out_ch, ch_mult, num_res_blocks, attn_resolutions, dropout, resamp_with_conv, in_channels,
-                 resolution, z_channels, give_pre_end=False, tanh_out=False, use_linear_attn=False,
-                 attn_type="vanilla"):
+    def __init__(self, ch, out_ch, ch_mult, num_res_blocks, attn_resolutions, dropout, resamp_with_conv,
+                 in_channels, resolution, z_channels, give_pre_end=False, tanh_out=False,
+                 use_linear_attn=False, attn_type="vanilla"):
         super().__init__()
         if use_linear_attn: attn_type = "linear"
         self.ch = ch
@@ -121,10 +119,10 @@ class Decoder(nn.Module):
         self.tanh_out = tanh_out
 
         # compute in_ch_mult, block_in and curr_res at lowest res
-        in_ch_mult = (1,)+tuple(ch_mult)
-        block_in = ch*ch_mult[self.num_resolutions-1]
-        curr_res = resolution // 2**(self.num_resolutions-1)
-        self.z_shape = (1,z_channels,curr_res,curr_res)
+        in_ch_mult = (1,) + tuple(ch_mult)
+        block_in = ch * ch_mult[self.num_resolutions - 1]
+        curr_res = resolution // 2 ** (self.num_resolutions - 1)
+        self.z_shape = (1, z_channels, curr_res, curr_res)
         print("Working with z of shape {} = {} dimensions.".format(
             self.z_shape, np.prod(self.z_shape)))
 
@@ -152,8 +150,8 @@ class Decoder(nn.Module):
         for i_level in reversed(range(self.num_resolutions)):
             block = nn.ModuleList()
             attn = nn.ModuleList()
-            block_out = ch*ch_mult[i_level]
-            for i_block in range(self.num_res_blocks+1):
+            block_out = ch * ch_mult[i_level]
+            for i_block in range(self.num_res_blocks + 1):
                 block.append(ResnetBlock(in_channels=block_in,
                                          out_channels=block_out,
                                          temb_channels=self.temb_ch,
@@ -167,7 +165,7 @@ class Decoder(nn.Module):
             if i_level != 0:
                 up.upsample = Upsample(block_in, resamp_with_conv)
                 curr_res = curr_res * 2
-            self.up.insert(0, up) # prepend to get consistent order
+            self.up.insert(0, up)  # prepend to get consistent order
 
         # end
         self.norm_out = Normalize(block_in)
@@ -178,7 +176,7 @@ class Decoder(nn.Module):
                                         padding=1)
 
     def forward(self, z, mid_feat=False):
-        #assert z.shape[1:] == self.z_shape[1:]
+        # assert z.shape[1:] == self.z_shape[1:]
         code_decoder_output = []
         self.last_z_shape = z.shape
 
@@ -195,17 +193,17 @@ class Decoder(nn.Module):
 
         # upsampling
         for i_level in reversed(range(self.num_resolutions)):
-            for i_block in range(self.num_res_blocks+1):
+            for i_block in range(self.num_res_blocks + 1):
                 h = self.up[i_level].block[i_block](h, temb)
                 if len(self.up[i_level].attn) > 0:
                     h = self.up[i_level].attn[i_block](h)
 
             if i_level != 2:
                 code_decoder_output.append(h)
-                                          
+
             if i_level != 0:
                 h = self.up[i_level].upsample(h)
-                
+
         if mid_feat:
             return code_decoder_output
         # end
@@ -220,11 +218,6 @@ class Decoder(nn.Module):
         return h
 
 
-
-
-
-
-
 class DCNv2Pack(ModulatedDeformConvPack):
     """Modulated deformable conv for deformable alignment.
     Different from the official DCNv2Pack, which generates offsets and masks
@@ -236,16 +229,15 @@ class DCNv2Pack(ModulatedDeformConvPack):
 
     def forward(self, x, feat):
         out = self.conv_offset(feat)
-        out =out.to(torch.float32)
+        out = out.to(torch.float32)
         o1, o2, mask = torch.chunk(out, 3, dim=1)
-        
+
         offset = torch.cat((o1, o2), dim=1)
-        
+
         mask = torch.sigmoid(mask)
 
-        
         return modulated_deform_conv(x, offset, mask, self.weight, self.bias, self.stride, self.padding,
-                                         self.dilation, self.groups, self.deformable_groups)
+                                     self.dilation, self.groups, self.deformable_groups)
 
 
 class WarpBlock(nn.Module):
@@ -256,16 +248,16 @@ class WarpBlock(nn.Module):
 
     def forward(self, x_vq, x_residual):
         x_residual = self.offset(torch.cat([x_vq, x_residual], dim=1))
-        
+
         feat_after_warp = self.dcn(x_vq, x_residual)
 
         return feat_after_warp
 
 
 class Decoder2(nn.Module):
-    def __init__(self, ch=128, out_ch=3, ch_mult=[1, 2, 4], num_res_blocks=2, attn_resolutions=[], dropout=0.0, resamp_with_conv=True, in_channels=3,
-                 resolution=256, z_channels=3, give_pre_end=False, tanh_out=False, use_linear_attn=False,
-                 attn_type="vanilla"):
+    def __init__(self, ch=128, out_ch=3, ch_mult=[1, 2, 4], num_res_blocks=2, attn_resolutions=[],
+                 dropout=0.0, resamp_with_conv=True, in_channels=3, resolution=256, z_channels=3,
+                 give_pre_end=False, tanh_out=False, use_linear_attn=False, attn_type="vanilla"):
         super().__init__()
         if use_linear_attn: attn_type = "linear"
         self.ch = ch
@@ -278,10 +270,10 @@ class Decoder2(nn.Module):
         self.tanh_out = tanh_out
 
         # compute in_ch_mult, block_in and curr_res at lowest res
-        in_ch_mult = (1,)+tuple(ch_mult)
-        block_in = ch*ch_mult[self.num_resolutions-1]
-        curr_res = resolution // 2**(self.num_resolutions-1)
-        self.z_shape = (1,z_channels,curr_res,curr_res)
+        in_ch_mult = (1,) + tuple(ch_mult)
+        block_in = ch * ch_mult[self.num_resolutions - 1]
+        curr_res = resolution // 2 ** (self.num_resolutions - 1)
+        self.z_shape = (1, z_channels, curr_res, curr_res)
         print("Working with z of shape {} = {} dimensions.".format(
             self.z_shape, np.prod(self.z_shape)))
 
@@ -309,8 +301,8 @@ class Decoder2(nn.Module):
         for i_level in reversed(range(self.num_resolutions)):
             block = nn.ModuleList()
             attn = nn.ModuleList()
-            block_out = ch*ch_mult[i_level]
-            for i_block in range(self.num_res_blocks+1):
+            block_out = ch * ch_mult[i_level]
+            for i_block in range(self.num_res_blocks + 1):
                 block.append(ResnetBlock(in_channels=block_in,
                                          out_channels=block_out,
                                          temb_channels=self.temb_ch,
@@ -324,7 +316,7 @@ class Decoder2(nn.Module):
             if i_level != 0:
                 up.upsample = Upsample(block_in, resamp_with_conv)
                 curr_res = curr_res * 2
-            self.up.insert(0, up) # prepend to get consistent order
+            self.up.insert(0, up)  # prepend to get consistent order
 
         # end
         self.norm_out = Normalize(block_in)
@@ -333,18 +325,16 @@ class Decoder2(nn.Module):
                                         kernel_size=3,
                                         stride=1,
                                         padding=1)
-        self.warp = nn.ModuleList()   
+        self.warp = nn.ModuleList()
         self.warp.append(WarpBlock(256))
         self.warp.append(WarpBlock(128))
-
 
         self.mix = nn.ModuleList()
         self.mix.append(Mix(m=-1.0))
         self.mix.append(Mix(m=-0.6))
 
-
     def forward(self, z, code_decoder_output, enc_feat):
-        #assert z.shape[1:] == self.z_shape[1:]
+        # assert z.shape[1:] == self.z_shape[1:]
         self.last_z_shape = z.shape
 
         # timestep embedding
@@ -360,24 +350,20 @@ class Decoder2(nn.Module):
 
         # upsampling
         for i_level in reversed(range(self.num_resolutions)):
-            for i_block in range(self.num_res_blocks+1):
+            for i_block in range(self.num_res_blocks + 1):
                 h = self.up[i_level].block[i_block](h, temb)
                 if len(self.up[i_level].attn) > 0:
                     h = self.up[i_level].attn[i_block](h)
 
-            if i_level!=2:
-                #print(i_level)
-                code_decoder_output[1-i_level]=code_decoder_output[1-i_level].to('cuda')
-
-                code_decoder_output[1-i_level]=code_decoder_output[1-i_level].to(torch.float32)
+            if i_level != 2:
+                # print(i_level)
+                code_decoder_output[1 - i_level] = code_decoder_output[1 - i_level].to('cuda')
+                code_decoder_output[1 - i_level] = code_decoder_output[1 - i_level].to(torch.float32)
                 h = h.to(torch.float32)
-                h = self.mix[1-i_level](enc_feat[i_level], h)
+                h = self.mix[1 - i_level](enc_feat[i_level], h)
 
-                x_vq = self.warp[1-i_level](code_decoder_output[1-i_level], h)
+                x_vq = self.warp[1 - i_level](code_decoder_output[1 - i_level], h)
                 h = h + x_vq * (h.mean() / x_vq.mean())
-
-
-
 
             if i_level != 0:
                 h = self.up[i_level].upsample(h)
@@ -392,7 +378,7 @@ class Decoder2(nn.Module):
         if self.tanh_out:
             h = torch.tanh(h)
         return h
-    
+
 
 class Mix(nn.Module):
     def __init__(self, m=-0.80):
